chat: factor out generate unit tests button
Also added an empty transcript fixture so we can set the correct
arguments on storybook to see the button.

Test Plan: http://localhost:6007/?path=/story/cody-chat--default&args=transcript:empty;experimentalUnitTestEnabled:!true
